refactor julia_con into a package as well
